+++ verifier/src/main/java/com/aws/jverify/verifier/Driver.java
@@ -2,7 +2,7 @@ package com.aws.jverify.verifier;
 
 import com.aws.jverify.common.Position;
 import com.aws.jverify.verifier.compiler.JavaToDafnyCompiler;
-import com.aws.jverify.verifier.compiler.desugar.java.NameCompiler;
+import com.aws.jverify.verifier.compiler.transformations.NameCompiler;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
