+++ verifier/src/main/java/module-info.java
@@ -13,6 +13,6 @@ module com.aws.jverify.verifier {
     requires org.jgrapht.core;
 
     opens com.aws.jverify.verifier to info.picocli;
-    opens com.aws.jverify.verifier.compiler.desugar.java to info.picocli;
+    opens com.aws.jverify.verifier.compiler.transformations to info.picocli;
     opens com.aws.jverify.verifier.compiler to info.picocli;
 }
